feat(api): move create/close window in GUI
