+++ Holovibes/includes/API.hh
@@ -223,15 +223,6 @@ void set_record_queue_location(Device device);
  */
 void set_record_buffer_size(uint value);
 
-void set_light_ui_mode(bool value);
-
-bool is_light_ui_mode();
-
-/*! \brief Closes all the currently displaying windows
- *
- */
-void close_windows();
-
 /*! \brief Set the camera timeout object */
 void set_camera_timeout();
 
@@ -255,13 +246,6 @@ void enable_pipe_refresh();
  *
  */
 void disable_pipe_refresh();
-
-/*! \brief Create and open a window of the specified size and kind
- *
- * \param[in] window_kind the kind of window to create (raw or holographic window)
- * \param[in] window_size the size of the window
- */
-void create_window(Computation window_kind, ushort window_size);
 
 void create_pipe();
