+++ Holovibes/sources/gui/GUI.cc
@@ -10,6 +10,104 @@
 namespace holovibes::gui
 {
 
+void set_light_ui_mode(bool value)
+{
+    auto path = holovibes::settings::user_settings_filepath;
+    std::ifstream input_file(path);
+    json j_us = json::parse(input_file);
+    j_us["light_ui"] = value;
+
+    std::ofstream output_file(path);
+    output_file << j_us.dump(1);
+}
+
+bool is_light_ui_mode()
+{
+    auto path = holovibes::settings::user_settings_filepath;
+    std::ifstream input_file(path);
+    json j_us = json::parse(input_file);
+
+    return json_get_or_default(j_us, false, "light_ui");
+}
+
+QPoint getSavedHoloWindowPos()
+{
+    auto path = holovibes::settings::user_settings_filepath;
+    std::ifstream input_file(path);
+    json j_us = json::parse(input_file);
+
+    int x = json_get_or_default(j_us, 0, "holo window", "x");
+    int y = json_get_or_default(j_us, 0, "holo window", "y");
+    return QPoint(x, y);
+}
+
+QSize getSavedHoloWindowSize(ushort& width, ushort& height)
+{
+    auto path = holovibes::settings::user_settings_filepath;
+    std::ifstream input_file(path);
+    json j_us = json::parse(input_file);
+
+    int final_width = json_get_or_default(j_us, width, "holo window", "width");
+    int final_height = json_get_or_default(j_us, height, "holo window", "height");
+    return QSize(final_width, final_height);
+}
+
+void close_windows()
+{
+    if (UserInterfaceDescriptor::instance().mainDisplay.get() != nullptr)
+        UserInterfaceDescriptor::instance().mainDisplay.get()->save_gui("holo window");
+    UserInterfaceDescriptor::instance().mainDisplay.reset(nullptr);
+
+    UserInterfaceDescriptor::instance().sliceXZ.reset(nullptr);
+    UserInterfaceDescriptor::instance().sliceYZ.reset(nullptr);
+    UserInterfaceDescriptor::instance().filter2d_window.reset(nullptr);
+    UserInterfaceDescriptor::instance().lens_window.reset(nullptr);
+    UserInterfaceDescriptor::instance().plot_window_.reset(nullptr);
+    UserInterfaceDescriptor::instance().raw_window.reset(nullptr);
+}
+
+void create_window(Computation window_kind, ushort window_size)
+{
+    const camera::FrameDescriptor& fd = api::get_fd();
+    unsigned short width = fd.width;
+    unsigned short height = fd.height;
+    get_good_size(width, height, window_size);
+
+    QPoint pos = getSavedHoloWindowPos();
+    QSize size = getSavedHoloWindowSize(width, height);
+
+    if (UI.mainDisplay)
+    {
+        pos = UI.mainDisplay->framePosition();
+        size = UI.mainDisplay->size();
+        UI.mainDisplay.reset(nullptr);
+    }
+
+    if (window_kind == Computation::Raw)
+    {
+        UI.mainDisplay.reset(new holovibes::gui::RawWindow(pos,
+                                                           size,
+                                                           api::get_input_queue().get(),
+                                                           static_cast<float>(width) / static_cast<float>(height)));
+        UI.mainDisplay->setBitshift(api::get_raw_bitshift());
+    }
+    else
+    {
+        UI.mainDisplay.reset(new gui::HoloWindow(pos,
+                                                 size,
+                                                 api::get_gpu_output_queue().get(),
+                                                 UI.sliceXZ,
+                                                 UI.sliceYZ,
+                                                 static_cast<float>(width) / static_cast<float>(height)));
+        UI.mainDisplay->set_is_resize(false);
+        UI.mainDisplay->resetTransform();
+        UI.mainDisplay->setAngle(api::get_rotation());
+        UI.mainDisplay->setFlip(api::get_horizontal_flip());
+    }
+
+    UI.mainDisplay->setTitle(QString("XY view"));
+}
+
 void set_filter2d_view(bool enabled, uint auxiliary_window_max_size)
 {
     if (enabled)
@@ -141,6 +239,14 @@ void active_noise_zone() { UI.mainDisplay->getOverlayManager().enable<gui::Noise
 
 void active_signal_zone() { UI.mainDisplay->getOverlayManager().enable<gui::Signal>(); }
 
+void set_reticle_overlay_visible(bool value)
+{
+    if (value)
+        UI.mainDisplay->getOverlayManager().enable<gui::Reticle>(false);
+    else
+        UI.mainDisplay->getOverlayManager().disable(gui::Reticle);
+}
+
 void open_advanced_settings(QMainWindow* parent, std::function<void()> callback)
 {
     UI.advanced_settings_window_ = std::make_unique<gui::AdvancedSettingsWindow>(parent);
@@ -176,8 +282,8 @@ const std::string browse_record_output_file(std::string& std_filepath)
 
     // Setting values in UserInterfaceDescriptor instance in a more optimized manner
     std::replace(parentPath.begin(), parentPath.end(), '/', '\\');
-    UserInterfaceDescriptor::instance().record_output_directory_ = std::move(parentPath);
-    UserInterfaceDescriptor::instance().output_filename_ = std::move(fileNameWithoutExt);
+    UI.record_output_directory_ = std::move(parentPath);
+    UI.output_filename_ = std::move(fileNameWithoutExt);
 
     return fileExt;
 }
