+++ Holovibes/sources/gui/windows/panels/image_rendering_panel.cc
@@ -135,9 +135,9 @@ void ImageRenderingPanel::set_computation_mode(int mode)
 
     Computation comp_mode = static_cast<Computation>(mode);
 
-    api::close_windows();
+    gui::close_windows();
     api::set_computation_mode(comp_mode);
-    api::create_window(comp_mode, parent_->window_max_size);
+    gui::create_window(comp_mode, parent_->window_max_size);
 
     if (comp_mode == Computation::Hologram)
     {
